weather_source/requirements.txt
fastapi
uvicorn
redis
celery
requests
python-dotenv
pymongo
kafka-python